# services/engine/openapi.json
{
  "components": {
    "schemas": {
      "AuditVerification": {
        "properties": {
          "broken_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broken Seq"
          },
          "checked": {
            "title": "Checked",
            "type": "integer"
          },
          "head_hash": {
            "title": "Head Hash",
            "type": "string"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "checked",
          "broken_seq",
          "head_hash"
        ],
        "title": "AuditVerification",
        "type": "object"
      },
      "ClockState": {
        "properties": {
          "running": {
            "title": "Running",
            "type": "boolean"
          },
          "sim_now": {
            "format": "date-time",
            "title": "Sim Now",
            "type": "string"
          },
          "speed": {
            "title": "Speed",
            "type": "number"
          },
          "wall_time_compression": {
            "title": "Wall Time Compression",
            "type": "number"
          }
        },
        "required": [
          "sim_now",
          "speed",
          "running",
          "wall_time_compression"
        ],
        "title": "ClockState",
        "type": "object"
      },
      "ClockUpdate": {
        "properties": {
          "advance_seconds": {
            "default": 0,
            "maximum": 31536000.0,
            "minimum": 0.0,
            "title": "Advance Seconds",
            "type": "number"
          },
          "running": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Running"
          },
          "sim_now": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sim Now"
          },
          "speed": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          }
        },
        "title": "ClockUpdate",
        "type": "object"
      },
...
      "DemoLoginRequest": {
        "properties": {
          "persona_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona Id"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "role"
        ],
        "title": "DemoLoginRequest",
        "type": "object"
      },
      "EntraCallbackResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "oidc_stub": {
            "default": true,
            "title": "Oidc Stub",
            "type": "boolean"
          },
          "principal": {
            "$ref": "#/components/schemas/Principal"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "principal"
        ],
        "title": "EntraCallbackResponse",
        "type": "object"
      },
      "GrantToken": {
        "properties": {
          "contact_note_due_at": {
            "format": "date-time",
            "title": "Contact Note Due At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expires_at",
          "contact_note_due_at"
        ],
        "title": "GrantToken",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "principal": {
            "$ref": "#/components/schemas/Principal"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "principal"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "PasskeyOptionsRequest": {
        "properties": {
          "persona_id": {
            "title": "Persona Id",
            "type": "string"
          }
        },
        "required": [
          "persona_id"
        ],
        "title": "PasskeyOptionsRequest",
        "type": "object"
      },
      "PasskeyOptionsResponse": {
        "properties": {
          "options": {
            "additionalProperties": true,
            "title": "Options",
            "type": "object"
          },
          "transaction_id": {
            "title": "Transaction Id",
            "type": "string"
          }
        },
        "required": [
          "transaction_id",
          "options"
        ],
        "title": "PasskeyOptionsResponse",
        "type": "object"
      },
      "PasskeyVerifyRequest": {
        "properties": {
          "credential": {
            "additionalProperties": true,
            "title": "Credential",
            "type": "object"
          },
          "transaction_id": {
            "title": "Transaction Id",
            "type": "string"
          }
        },
        "required": [
          "transaction_id",
          "credential"
        ],
        "title": "PasskeyVerifyRequest",
        "type": "object"
      },
      "Principal": {
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "scope_path": {
            "title": "Scope Path",
            "type": "string"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array",
            "uniqueItems": true
          },
          "subject_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Token"
          },
          "synthetic": {
            "default": true,
            "title": "Synthetic",
            "type": "boolean"
          }
        },
        "required": [
          "actor_id",
          "role",
          "scopes",
          "scope_path"
        ],
        "title": "Principal",
        "type": "object"
      },
      "Role": {
        "enum": [
          "personnel",
          "uwo",
          "counsellor",
          "mo",
          "commander",
          "hq",
          "wdec",
          "dpo",
          "hrms_integrator",
          "admin",
          "director"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WelfareGrantBody": {
        "properties": {
          "justification": {
            "maxLength": 500,
            "minLength": 8,
            "title": "Justification",
            "type": "string"
          },
          "purpose_code": {
            "maxLength": 64,
            "minLength": 2,
            "title": "Purpose Code",
            "type": "string"
          },
          "token": {
            "pattern": "^st_[a-z2-7]{16}$",
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "purpose_code",
          "justification"
        ],
        "title": "WelfareGrantBody",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Token-only foundation API",
    "title": "MANOBAL engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/demo-login": {
      "post": {
        "operationId": "demo_login_api_v1_auth_demo_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Login"
      }
    },
    "/api/v1/auth/entra/callback": {
      "get": {
        "operationId": "entra_callback_api_v1_auth_entra_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "minLength": 4,
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "app_role",
            "required": true,
            "schema": {
              "title": "App Role",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntraCallbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Entra Callback"
      }
    },
    "/api/v1/auth/passkey/login/options": {
      "post": {
        "operationId": "passkey_login_options_api_v1_auth_passkey_login_options_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyOptionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Passkey Login Options"
      }
    },
    "/api/v1/auth/passkey/login/verify": {
      "post": {
        "operationId": "passkey_login_verify_api_v1_auth_passkey_login_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Passkey Login Verify"
      }
    },
    "/api/v1/auth/passkey/register/options": {
      "post": {
        "operationId": "passkey_register_options_api_v1_auth_passkey_register_options_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyOptionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Passkey Register Options"
      }
    },
    "/api/v1/auth/passkey/register/verify": {
      "post": {
        "operationId": "passkey_register_verify_api_v1_auth_passkey_register_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasskeyVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Passkey Register Verify"
      }
    },
    "/api/v1/demo/clock": {
      "get": {
        "operationId": "read_demo_clock_api_v1_demo_clock_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClockState"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Demo Clock"
      },
      "post": {
        "operationId": "change_demo_clock_api_v1_demo_clock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClockUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClockState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Demo Clock"
      }
    },
    "/api/v1/gov/audit/verify": {
      "post": {
        "operationId": "audit_verify_api_v1_gov_audit_verify_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerification"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit Verify"
      }
    },
...
    "/api/v1/welfare/cases/{case_id}/grant": {
      "post": {
        "operationId": "create_case_grant_api_v1_welfare_cases__case_id__grant_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "pattern": "^MB-[0-9]{4}$",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WelfareGrantBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantToken"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Case Grant"
      }
    }
  }
}